+++ Magic/CHANGELOG.md
@@ -10,6 +10,7 @@
  - Improved support for using attributes in base spell parameters (cooldown, range, etc will show correctly in lore)
  - Fixed graduate spell list in potter configs
  - Cars in war configs can now go up 1 block slopes
+ - Add Growth spell (applies bonemeal in a radius)
 
 # 7.5